Reorganize DB files
